// src/generated/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  BigInt: { input: any; output: any; }
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf). */
  JSON: { input: any; output: any; }
};

export type Meta = {
  status: Maybe<Scalars['JSON']['output']>;
};

export type PageInfo = {
  endCursor: Maybe<Scalars['String']['output']>;
  hasNextPage: Scalars['Boolean']['output'];
  hasPreviousPage: Scalars['Boolean']['output'];
  startCursor: Maybe<Scalars['String']['output']>;
};
// ...
export type ActivityEventPage = {
  items: Array<ActivityEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type AddNftTierEvent = {
  caller: Scalars['String']['output'];
  category: Scalars['Int']['output'];
  chainId: Scalars['Int']['output'];
  encodedIpfsUri: Scalars['String']['output'];
  from: Scalars['String']['output'];
  hook: Scalars['String']['output'];
  id: Scalars['String']['output'];
  initialSupply: Scalars['Int']['output'];
  logIndex: Scalars['Int']['output'];
  metadata: Maybe<Scalars['JSON']['output']>;
  price: Scalars['BigInt']['output'];
  project: Maybe<Project>;
  projectId: Scalars['Int']['output'];
  remainingSupply: Scalars['Int']['output'];
  resolvedUri: Maybe<Scalars['String']['output']>;
  suckerGroupId: Scalars['String']['output'];
  tier: Maybe<NftTier>;
  tierId: Scalars['Int']['output'];
  timestamp: Scalars['Int']['output'];
  txHash: Scalars['String']['output'];
  version: Scalars['Int']['output'];
};
// ...
export type AddNftTierEventPage = {
  items: Array<AddNftTierEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type AddToBalanceEventPage = {
  items: Array<AddToBalanceEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type AutoIssueEventPage = {
  items: Array<AutoIssueEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type BorrowLoanEventPage = {
  items: Array<BorrowLoanEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type BridgeClaimEventPage = {
  items: Array<BridgeClaimEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type BurnEventPage = {
  items: Array<BurnEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type BuybackPoolEventPage = {
  items: Array<BuybackPoolEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type CashOutTokensEventPage = {
  items: Array<CashOutTokensEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type DeployErc20EventPage = {
  items: Array<DeployErc20Event>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type LiquidateLoanEventPage = {
  items: Array<LiquidateLoanEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type ManualBurnEventPage = {
  items: Array<ManualBurnEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type ManualMintTokensEventPage = {
  items: Array<ManualMintTokensEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type MintNftEventPage = {
  items: Array<MintNftEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type MintTokensEventPage = {
  items: Array<MintTokensEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type NftHook = {
  address: Scalars['String']['output'];
  chainId: Scalars['Int']['output'];
  createdAt: Scalars['Int']['output'];
  name: Maybe<Scalars['String']['output']>;
  nftTiers: Maybe<NftTierPage>;
  nfts: Maybe<NftPage>;
  project: Maybe<Project>;
  projectId: Scalars['Int']['output'];
  symbol: Maybe<Scalars['String']['output']>;
  version: Scalars['Int']['output'];
};
// ...
export type NftHookPage = {
  items: Array<NftHook>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};

export type NftPage = {
  items: Array<Nft>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type OperatorPermissionsSetEventPage = {
  items: Array<OperatorPermissionsSetEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type ParticipantPage = {
  items: Array<Participant>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type PayEventPage = {
  items: Array<PayEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type PermissionHolderPage = {
  items: Array<PermissionHolder>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};

export type Project = {
  activityEvents: Maybe<ActivityEventPage>;
  addNftTierEvents: Maybe<AddNftTierEventPage>;
  addToBalanceEvents: Maybe<AddToBalanceEventPage>;
  autoIssueEvents: Maybe<AutoIssueEventPage>;
  balance: Scalars['BigInt']['output'];
  borrowLoanEvents: Maybe<BorrowLoanEventPage>;
  bridgeClaimEvents: Maybe<BridgeClaimEventPage>;
  burnEvents: Maybe<BurnEventPage>;
  buybackPoolEvents: Maybe<BuybackPoolEventPage>;
  cashOutTokensEvents: Maybe<CashOutTokensEventPage>;
  chainId: Scalars['Int']['output'];
  contributorsCount: Scalars['Int']['output'];
  coverImageUri: Maybe<Scalars['String']['output']>;
  createdAt: Scalars['Int']['output'];
  createdWithinTrendingWindow: Maybe<Scalars['Boolean']['output']>;
  creator: Scalars['String']['output'];
  currency: Maybe<Scalars['BigInt']['output']>;
  decimals: Maybe<Scalars['Int']['output']>;
  deployErc20Events: Maybe<DeployErc20EventPage>;
  deployer: Scalars['String']['output'];
  description: Maybe<Scalars['String']['output']>;
  discord: Maybe<Scalars['String']['output']>;
  domain: Maybe<Scalars['String']['output']>;
  farcaster: Maybe<Scalars['String']['output']>;
  handle: Maybe<Scalars['String']['output']>;
  id: Scalars['String']['output'];
  infoUri: Maybe<Scalars['String']['output']>;
  isRevnet: Maybe<Scalars['Boolean']['output']>;
  liquidateLoanEvents: Maybe<LiquidateLoanEventPage>;
  logoUri: Maybe<Scalars['String']['output']>;
  manualBurnEvents: Maybe<ManualBurnEventPage>;
  manualMintTokensEvents: Maybe<ManualMintTokensEventPage>;
  metadata: Maybe<Scalars['JSON']['output']>;
  metadataUri: Maybe<Scalars['String']['output']>;
  mintNftEvents: Maybe<MintNftEventPage>;
  mintTokensEvents: Maybe<MintTokensEventPage>;
  name: Maybe<Scalars['String']['output']>;
  nftHooks: Maybe<NftHookPage>;
  nfts: Maybe<NftPage>;
  nftsMintedCount: Scalars['Int']['output'];
  operatorPermissionsSetEvents: Maybe<OperatorPermissionsSetEventPage>;
  owner: Scalars['String']['output'];
  participants: Maybe<ParticipantPage>;
  payDisclosure: Maybe<Scalars['String']['output']>;
  payEvents: Maybe<PayEventPage>;
  paymentsCount: Scalars['Int']['output'];
  permissionHolders: Maybe<PermissionHolderPage>;
  projectCreateEvents: Maybe<ProjectCreateEventPage>;
  projectId: Scalars['Int']['output'];
  projectMoments: Maybe<ProjectMomentPage>;
  projectPayers: Maybe<ProjectPayerPage>;
  projectTagline: Maybe<Scalars['String']['output']>;
  projectTransferEvents: Maybe<ProjectTransferEventPage>;
  reallocateLoanEvents: Maybe<ReallocateLoanEventPage>;
  redeemCount: Scalars['Int']['output'];
  redeemVolume: Scalars['BigInt']['output'];
  redeemVolumeUsd: Scalars['BigInt']['output'];
  removeNftTierEvents: Maybe<RemoveNftTierEventPage>;
  repayLoanEvents: Maybe<RepayLoanEventPage>;
  reservedTokenSupply: Scalars['BigInt']['output'];
  rulesetQueuedEvents: Maybe<RulesetQueuedEventPage>;
  sendPayoutToSplitEvents: Maybe<SendPayoutToSplitEventPage>;
  sendPayoutsEvents: Maybe<SendPayoutsEventPage>;
  sendReservedTokensToSplitEvents: Maybe<SendReservedTokensToSplitEventPage>;
  sendReservedTokensToSplitsEvents: Maybe<SendReservedTokensToSplitsEventPage>;
  setUriEvents: Maybe<SetUriEventPage>;
  storeAutoIssuanceAmountEvent: Maybe<StoreAutoIssuanceAmountEventPage>;
  suckerGroup: Maybe<SuckerGroup>;
  suckerGroupId: Scalars['String']['output'];
  swapEvents: Maybe<SwapEventPage>;
  tags: Maybe<Array<Scalars['String']['output']>>;
  telegram: Maybe<Scalars['String']['output']>;
  token: Maybe<Scalars['String']['output']>;
  tokenSupply: Scalars['BigInt']['output'];
  tokenSymbol: Maybe<Scalars['String']['output']>;
  tokens: Maybe<Array<Scalars['String']['output']>>;
  trendingPaymentsCount: Scalars['Int']['output'];
  trendingScore: Scalars['BigInt']['output'];
  trendingVolume: Scalars['BigInt']['output'];
  trendingVolumeUsd: Scalars['BigInt']['output'];
  twitter: Maybe<Scalars['String']['output']>;
  useAllowanceEvents: Maybe<UseAllowanceEventPage>;
  version: Scalars['Int']['output'];
  volume: Scalars['BigInt']['output'];
  volumeUsd: Scalars['BigInt']['output'];
};
// ...
export type ProjectCreateEventPage = {
  items: Array<ProjectCreateEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type ProjectMomentPage = {
  items: Array<ProjectMoment>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type ProjectPayerPage = {
  items: Array<ProjectPayer>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type ProjectTransferEventPage = {
  items: Array<ProjectTransferEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type ReallocateLoanEventPage = {
  items: Array<ReallocateLoanEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type RemoveNftTierEventPage = {
  items: Array<RemoveNftTierEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type RepayLoanEventPage = {
  items: Array<RepayLoanEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type RulesetQueuedEventPage = {
  items: Array<RulesetQueuedEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type SendPayoutToSplitEventPage = {
  items: Array<SendPayoutToSplitEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type SendPayoutsEventPage = {
  items: Array<SendPayoutsEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type SendReservedTokensToSplitEventPage = {
  items: Array<SendReservedTokensToSplitEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type SendReservedTokensToSplitsEventPage = {
  items: Array<SendReservedTokensToSplitsEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type SetUriEventPage = {
  items: Array<SetUriEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type StoreAutoIssuanceAmountEventPage = {
  items: Array<StoreAutoIssuanceAmountEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};

export type SuckerGroup = {
  addresses: Array<Scalars['String']['output']>;
  balance: Scalars['BigInt']['output'];
  contributorsCount: Scalars['Int']['output'];
  createdAt: Scalars['Int']['output'];
  id: Scalars['String']['output'];
  nftsMintedCount: Scalars['Int']['output'];
  paymentsCount: Scalars['Int']['output'];
  projects: Maybe<ProjectPage>;
  redeemCount: Scalars['Int']['output'];
  redeemVolume: Scalars['BigInt']['output'];
  redeemVolumeUsd: Scalars['BigInt']['output'];
  reservedTokenSupply: Scalars['BigInt']['output'];
  suckerTransactions: Maybe<SuckerTransactionPage>;
  tokenSupply: Scalars['BigInt']['output'];
  trendingPaymentsCount: Scalars['Int']['output'];
  trendingScore: Scalars['BigInt']['output'];
  trendingVolume: Scalars['BigInt']['output'];
  version: Scalars['Int']['output'];
  volume: Scalars['BigInt']['output'];
  volumeUsd: Scalars['BigInt']['output'];
};
// ...
export type SwapEventPage = {
  items: Array<SwapEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};
// ...
export type UseAllowanceEventPage = {
  items: Array<UseAllowanceEvent>;
  pageInfo: PageInfo;
  totalCount: Scalars['Int']['output'];
};